requirements.txt
# Hyperlocal News Summarizer - Dependencies

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Feed Parsing
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2

# NLP & ML
transformers>=4.35.0
torch>=2.2.0
sentencepiece>=0.1.99
protobuf>=4.25.0

# Translation
googletrans==4.0.0rc1
indic-transliteration==1.9.2

# NER & Text Processing
spacy==3.7.2
nltk==3.8.1

# Geo-tagging
geopy==2.4.1
geocoder==1.38.1

# Voice Synthesis
gTTS==2.4.0
pyttsx3==2.90

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Benchmarking & Evaluation
evaluate>=0.4.0
rouge-score>=0.1.2

